Removes the assigment of generated AR::Base classes to constants.
This functionality removes the potential for hidden coupling between
Vorpal::Engines. The caveat now is that the AR::Base classes are no
longer cached outside of the Vorpal::Engine. This means that every time
a Vorpal::Engine is created, new AR::Base classes will also be created
unless they are specified by the client code (using the :to option on
the #map method.)