replaces .ruby-version with a min Ruby version in the gemspec
It seemed odd for a Ruby library to set a specific Ruby version to develop again. It was also especially concerning that an old, unsupported Ruby version was being used. Instead, it should be safer to allow the developer to choose any version greater than a minimum. And since the last stable version of Vorpal was developed against 2.1.6, that thus felt like a safe minimum to use.